% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exper_pred.R
\name{exper_pred}
\alias{exper_pred}
\title{Simulate an experiment using ODE model}
\usage{
exper_pred(
  parms,
  x0,
  ode_fun,
  transf.num,
  transf.dur = 48,
  transf.dil = 5/200,
  aug.ls = list(),
  reso = 10,
  return.all = TRUE
)
}
\arguments{
\item{parms}{parameters for model implemented between transfers (this is first because \code{optim()} demands this)}

\item{x0}{vector of initial conditions at start of transfer period 1 (e.g. at transfer 0)}

\item{ode_fun}{The ode function for model implemented between transfers.}

\item{transf.num}{How many transfers to model (1 is basically no dealing with transfers). With Landis data, this will generally be 6.}

\item{transf.dur}{How many hours are each transfer period? Default is 48 (e.g. Landis data).}

\item{transf.dil}{At the beginning of each transfer period, what is the fraction of innoculant in the total material. Default is 5/200 (from Landis: 5 microliters innoculant into 195 new flour goo).}

\item{aug.ls}{Augmentation list. Currently used to identify modification for Tilman model (see \code{ode_Til()} and \code{ode_Til_1spec()}). Argument \emph{$Til} should be a numeric for the number of species, and \emph{$TilR$} is the amount of resouces at the start of each transfer (arbitrary, defaults to 195 for the microliters in Landis 2021).}

\item{reso}{How many time points per hour should we return for plotting?}

\item{return.all}{Should we return both "highly" resolved trajectories for plotting (resolution based on \code{reso}) and predictions at each transfer (\code{TRUE}), or should we just return the predictions at each transfer (\code{FALSE}). \code{FALSE} provides everything needed for model-fitting, and should be faster (although how much is unclear).}
}
\value{
For \code{return.all=TRUE}, returns a list, with \verb{$transf.pred} being a data frame of predictions at each transfer point, and \verb{$series.tot} being a data frame of predictions every 1/\code{reso} hours. For \code{return.all=FALSE}, only returns the data frame of \code{transf.pred}.
}
\description{
This function takes parameters, model, initial conditions, and experiment design (number of transfers, dilution), and simulates trajectory of microbes.
}
