% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_helper.R
\name{obj_helper}
\alias{obj_helper}
\title{Apply objective function to multiple experimental replicates}
\usage{
obj_helper(
  parms,
  x0.mat,
  ode_fun,
  transf.num,
  dat.real.ls,
  transf.dur = 48,
  transf.dil = 5/200,
  aug.ls = list()
)
}
\arguments{
\item{parms}{parameters for model implemented between transfers (this is first because \code{optim()} demands this)}

\item{x0.mat}{matrix of initial conditions at start of transfer 1; each row is a different replicate. Note that this MUST be in the same order as dat.real.ls below}

\item{ode_fun}{The ode function for model implemented between transfers.}

\item{transf.num}{How many transfers to model (1 is basically no dealing with transfers). With Landis data, this will generally be 6.}

\item{dat.real.ls}{Empirical data to fit, in the form of a list. Each entry is a \emph{matrix}, with columns of replicate number ('rep'), transfer number ('transf'), and abundance for each species ('abund1', 'abund2', etc)>
\code{dataprep_landis()} preps the Landis data for this, and serves as an example.}

\item{transf.dur}{How many hours are each transfer period? Default is 48 (e.g. Landis data).}

\item{transf.dil}{At the beginning of each transfer period, what is the fraction of innoculant in the total material. Default is 5/200 (from Landis: 5 microliters innoculant into 195 new flour goo).}

\item{aug.ls}{Augmentation list. Not currently used, but a way to pass information down the chain of functions.}
}
\value{
Returns single value of the sum of the values from objective functions for each replicate.
}
\description{
This function applies \code{SourCoex::obj_ss()} to the supplied list of data sets and initial conditions (e.g. multiple replicates). As such it serves as an intermediate between \code{optim()} and \code{obj_ss()}.
}
\details{
Note: it may be appropriate to update this with multiple objective function options in the future.

Note: this is a function that should be updated when we switch over to using hours.

Also note: this function is quite fragile wrt the format of the real data. \strong{CBE should add some safety features and or more flexibility.}
}
