% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_ss.R
\name{obj_ss}
\alias{obj_ss}
\title{Objective function using sum of squares}
\usage{
obj_ss(
  parms,
  x0,
  ode_fun,
  transf.num,
  dat.real,
  transf.dur = 48,
  transf.dil = 5/200,
  aug.ls = list()
)
}
\arguments{
\item{parms}{parameters for model implemented between transfers (this is first because \code{optim()} demands this)}

\item{x0}{vector of initial conditions at start of transfer period 1 (e.g. at transfer 0)}

\item{ode_fun}{The ode function for model implemented between transfers.}

\item{transf.num}{How many transfers to model (1 is basically no dealing with transfers). With Landis data, this will generally be 6.}

\item{dat.real}{actual data, in the form of a matrix, with 1st column being transfer number, second being abund species 1, etc. Should not include initial transfer unless we're fitting initial conditions as parameters
\code{dataprep_landis()} preps the Landis data for this, and serves as an example.}

\item{transf.dur}{How many hours are each transfer period? Default is 48 (e.g. Landis data).}

\item{transf.dil}{At the beginning of each transfer period, what is the fraction of innoculant in the total material. Default is 5/200 (from Landis: 5 microliters innoculant into 195 new flour goo).}

\item{aug.ls}{Augmentation list. Not currently used, but a way to pass information down the chain of functions.}
}
\value{
Returns single value of the sum of squared error.
}
\description{
This function fits the data of a single experiment for given parameters (relying on \code{exper_pred()}, compares results with actual data, and returns sum of squared error.
}
\details{
Note that at present this function does NOT treat initial conditions as a parameter to fit.
}
