% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_diagnostics.R
\name{ode_diagnostics}
\alias{ode_diagnostics}
\title{Check ODE fit}
\usage{
ode_diagnostics(
  fit1,
  fit2,
  parms.guess,
  x0.mat,
  ode_fun,
  transf.num = 6,
  dat.real.ls,
  scale.vec,
  aug.ls = list()
)
}
\arguments{
\item{fit1}{first fitting using optim}

\item{fit2}{second fitting using optim}

\item{Need}{to fill in params}
}
\value{
Nothing. Will print diagnostics to console.
}
\description{
This function give useful diagnostics for how parameter values changed through fits, how objective function values changed, and if fits converged.
}
