% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_log.R
\name{ode_log}
\alias{ode_log}
\title{Logistic Growth (1-species)}
\usage{
ode_log(t, y, parms)
}
\arguments{
\item{t}{time (only relevant for non-autonomous ODEs, not for us, but necessary for \code{deSolve} to play nice.}

\item{y}{current abundance of the population}

\item{parms}{parameters for the ODE model. In order:
\itemize{
\item r - intrinsic growth rate
\item k - carrying capacity
}}
}
\value{
List of the derivate at current population density.
}
\description{
ODE model for logistic growth. Note that is mostly an internal file: it gets used by \code{lsoda} in \code{exper_pred()}, and we don't have to interact with it directly.
}
\details{
Note that the names of the parameters is provided in \code{parmnamesLog}, and the units of the parameters is provided in \code{unitsLog}. These are both used in plotting functions (to generate nice tables of parameters).
}
