% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_LV.R
\name{ode_LV}
\alias{ode_LV}
\title{Lotka Volterra competition (2-species)}
\usage{
ode_LV(t, y, parms)
}
\arguments{
\item{t}{time (only relevant for non-autonomous ODEs, not for us, but necessary for \code{deSolve} to play nice.}

\item{y}{vector of current abundance of species 1, species 2}

\item{parms}{parameters for the ODE model. In order:
\itemize{
\item r1 - intrinsic growth rate of species 1
\item r2 - intrinsic growth rate of species 2
\item a12 - competition coefficient for species 2 on species 1
\item a21 - competition coefficient for species 1 on species 2
\item k1 - carrying capacity for species 1
\item k2 - carrying capacity for species 2
}}
}
\value{
List of derivates at current population densities.
}
\description{
ODE model for L-V competition. Note that is mostly an internal file: it gets used by \code{lsoda} in \code{exper_pred()}, and we don't have to interact with it directly.
}
\details{
Note that the names of the parameters is provided in \code{parmnames.LV}, and the units of the parameters is provided in \code{units.LV}. These are both used in plotting functions (to generate nice tables of parameters).
}
