% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til_1spec.R
\name{ode_Til_1spec}
\alias{ode_Til_1spec}
\title{Tilman competition (1-species + resource)}
\usage{
ode_Til_1spec(t, y, parms)
}
\arguments{
\item{t}{time (only relevant for non-autonomous ODEs, not for us, but necessary for \code{deSolve} to play nice.}

\item{y}{vector of current abundance of species 1, species 2}

\item{parms}{parameters for the ODE model. In order:
\itemize{
\item r1 - intrinsic growth rate of species 1
\item d1 - per capita mortality rate, species 1
}}
}
\value{
List of derivates at current population densities.
}
\description{
ODE model for Tilman competition with inert medium, for a single species. Here we know that the resource isn't growing, so \emph{r} in the classic Tilman model is 0.
}
\details{
Note that I believe the amount of resources available at the start of a transfer can be set to some constant (doesn't need to be fit). This is handled in the exper_pred, but for simplicity I will set it to 195, for the mL of new resource.

\strong{IMPORTANT!!} For this model to work, need to pass \emph{aug.ls$Til = 1} to \code{exper_pred()} (e.g. through the chain of parent functions).
\emph{aug.ls$TilR} is the amount of resources provided at the start of each transfer. If undefined, defaults to 195 (e.g. Landis et al. 2021 numbersw).

Note that the names of the parameters is provided in \code{parmnamesTil1spec}, and the units of the parameters is provided in \code{unitsTil1spec}. These are both used in plotting functions (to generate nice tables of parameters).
}
