% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_check.R
\name{parm_check}
\alias{parm_check}
\title{Check parameters}
\usage{
parm_check(parms.cur, parm.lower, parm.upper, parmnames, tol = 0.001)
}
\arguments{
\item{parms.cur}{Current estimate (from \code{optim()} etc)}

\item{parm.lower}{Lower bound that user supplied for \code{optim()} call}

\item{parm.upper}{Upper bound that user supplied}

\item{parmnames}{vector of the parameter names. Function uses this to identify boundary parameters}

\item{tol}{Tolerance parameter. Smaller values mean only estimates extremely close to the boundary are identified. Default is 0.001}
}
\value{
Nothing. Will output warnings if there are boundary parameters
}
\description{
This function checks to see if fitted parameters are at or near the user-defined parameters (returns informative warnings if so). This is useful for identifying when solvers may be "caught" at the boundaries, or when boundaries are too constraining.
}
