% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til_1spec.R
\docType{data}
\name{parmnames.Til.1spec}
\alias{parmnames.Til.1spec}
\title{List of parameter names for the 1-species Tilman competition model}
\format{
A vector of 3 strings.
}
\usage{
parmnames.Til.1spec
}
\description{
A companion to \code{ode_Til_1spec}.
}
\seealso{
\code{ode_Til_1spec}
}
\keyword{datasets}
