% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_log.R
\docType{data}
\name{parmnamesLog}
\alias{parmnamesLog}
\title{List of parameter names for the logistic growth model}
\format{
A vector of 2 strings.
}
\usage{
parmnamesLog
}
\description{
A companion to \code{ode_log}.
}
\seealso{
\code{ode_log}
}
\keyword{datasets}
