% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til.R
\docType{data}
\name{parmnamesTil}
\alias{parmnamesTil}
\title{List of parameter names for the Tilman competition (R*) model}
\format{
A vector of 4 strings.
}
\usage{
parmnamesTil
}
\description{
A companion to \code{ode_Til}.
}
\seealso{
\code{ode_Til}
}
\keyword{datasets}
