% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter_landis.R
\name{plotter_landis}
\alias{plotter_landis}
\title{Plot fitted models and Landis data}
\usage{
plotter_landis(
  parms,
  ode_fun,
  parmnames,
  parmunits,
  dat.comp,
  dat.solo1,
  dat.solo2,
  specid,
  aug.ls = list(),
  specmap.cur = specMap,
  reso = 10
)
}
\arguments{
\item{parms}{vector of parameter values for model implemented between transfers (this is first because \code{optim()} demands this).}

\item{ode_fun}{function for the ode model model implemented between transfers.}

\item{parmnames}{vectofr of characters for the name of the parameters of the ode function. These are generally saved in the ode function files as parnames.XX, e.g. \code{parnames.LV}, \code{parnames.log}.}

\item{parmunits}{Vector of characters for the units of the parameters of the ode function. These are also generally saved in the ode function files, in the form of \code{units.LV}, \code{units.log}, etc).}

\item{dat.comp}{dataframe of the data from the interspecific experiments. Generated in the right format with \code{dataprep_landis()}.}

\item{dat.solo1}{dataframe of the data from the single-species experiment for species 1. Generated in the right format with \code{dataprep_landis()}.}

\item{dat.solo2}{dataframe of the data from the single-species experiment for species 2. Generated in the right format with \code{dataprep_landis()}.}

\item{specid}{vector of species names, for generating appropriate titles}

\item{specmap.cur}{map of species id to species name. For the landis data in \code{sourData}, the appropriate map is \code{specMap}.}

\item{reso}{How many time points per hour should we generate to plot trajectories}
}
\value{
Returns a ggplot2 object: 2x2 grid of plots and parameter table. Actually a 3x2 grid with row for custom title object.
}
\description{
This function takes (presumably fitted) parameter values, model and experiment information, and raw data, and provides a panel each for experiments of interspecies competition, each of the single-species competition experiments, and parameter values.
}
