% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sour.data}
\alias{sour.data}
\title{Sourdough Data}
\format{
A data frame with 1080 rows and 6 variables.
\describe{
\item{abund}{Are the values relative ("rel") or absolute ("abs") abundance?}
\item{transf}{What is the transfer number? Multiply by 48 for hours since experiment start}
\item{spec1}{Identifier of first competing species. See "spec.map" to get genus/species from the identifier}
\item{spec2}{Identifier of second competing species.}
\item{abund1}{relative or absolute abundance of species 1}
\item{abund2}{relative or absolute abundance of species 2}
}
}
\source{
\url{https://elifesciences.org/articles/61644}
}
\usage{
sour.data
}
\description{
A data frame of timeseries data from pairwise competition experiments of sourdough communities from Landis et al. 2020.
}
\details{
Note from the data file of Landis et al:
"CFU counts and relative abundance data from competitions, transfers 1, 3, and 6 - Pairs were transferred every 48 hours through filtered flour media (see materials and methods)."
}
\seealso{
spec.map
}
\keyword{datasets}
