% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourTest_SS.R
\name{SourTest_SS}
\alias{SourTest_SS}
\title{Test SS objective function}
\usage{
SourTest_SS(
  parms.0,
  x0.mat,
  ode_fun,
  transf.num,
  dat.real.ls,
  transf.dur = 48,
  transf.dil = 5/200,
  aug.ls = list()
)
}
\arguments{
\item{x0.mat}{matrix of initial conditions at start of transfer 1; each row is a different replicate. Note that this MUST be in the same order as dat.real.ls below}

\item{ode_fun}{The ode function for model implemented between transfers.}

\item{transf.num}{How many transfers to model (1 is basically no dealing with transfers). With Landis data, this will generally be 6.}

\item{dat.real.ls}{Empirical data to fit, in the form of a list. Each entry is a \emph{matrix}, with columns of replicate number ('rep'), transfer number ('transf'), and abundance for each species ('abund1', 'abund2', etc)>
\code{dataprep_landis()} preps the Landis data for this, and serves as an example.}

\item{transf.dur}{How many hours are each transfer period? Default is 48 (e.g. Landis data).}

\item{transf.dil}{At the beginning of each transfer period, what is the fraction of innoculant in the total material. Default is 5/200 (from Landis: 5 microliters innoculant into 195 new flour goo).}

\item{aug.ls}{Augmentation list. Used to carry additional information, like identifying the Tilman model.}

\item{parms0}{parameters for model implemented between transfers, WITH VALUES CHOSEN TO ENSURE 0 ABUNDANCE AT ALL MEASUREMENT PERIODS}
}
\value{
Returns vector with the SS calculated by \code{obj_helper()} (first entry, labeled "objectiveSS"), and calculated by hand assuming real answer is at 0 (second entry, sanitySS). If these match, all is well.
}
\description{
This function tests that obj_helper.R is behaving. Takes identical arguments to obj_helper, but parameters should specific values such that microbial abundance is always 0.
}
