% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec.map}
\alias{spec.map}
\title{Map of sourdough species}
\format{
A data frame with 8 rows and 2 variables.
\describe{
\item{name.data}{Identifier used in sour.data. Note that numbers correspond to fungi, letters to bacteria.}
\item{name.full}{Scientific name}
}
}
\source{
\url{https://elifesciences.org/articles/61644}
}
\usage{
spec.map
}
\description{
A companion to sour.data, matching species ID to the genus and species name.
}
\seealso{
sour.data
}
\keyword{datasets}
