% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_log.R
\docType{data}
\name{units.log}
\alias{units.log}
\title{List of parameter units for the logistic growth model}
\format{
A vector of 2 strings.
}
\usage{
units.log
}
\description{
A companion to \code{ode_log}.
}
\seealso{
\code{ode_log}
}
\keyword{datasets}
