% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til.R
\docType{data}
\name{units.Til}
\alias{units.Til}
\title{List of parameter units for the Tilman competition (R*) model}
\format{
A vector of 4 strings.
}
\usage{
units.Til
}
\description{
A companion to \code{ode_Til}.
}
\seealso{
\code{ode_Til}
}
\keyword{datasets}
