% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_LV.R
\docType{data}
\name{unitsLV}
\alias{unitsLV}
\title{List of parameter units for the Lotka Voltera competition model}
\format{
A vector of 6 strings.
}
\usage{
unitsLV
}
\description{
A companion to \code{ode_LV}.
}
\seealso{
\code{ode_LV}
}
\keyword{datasets}
