% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til.R
\docType{data}
\name{unitsTil}
\alias{unitsTil}
\title{List of parameter units for the Tilman competition (R*) model}
\format{
A vector of 4 strings.
}
\usage{
unitsTil
}
\description{
A companion to \code{ode_Til}.
}
\seealso{
\code{ode_Til}
}
\keyword{datasets}
