% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_Til_1spec.R
\docType{data}
\name{unitsTil1spec}
\alias{unitsTil1spec}
\title{List of parameter units for the 1-species Tilman model competition model}
\format{
A vector of 3 strings.
}
\usage{
unitsTil1spec
}
\description{
A companion to \code{ode_Til_1spec}.
}
\seealso{
\code{ode_Til_1spec}
}
\keyword{datasets}
